FIX - Clean up username logging.
Change in functionality: the bot now logs the most recent 20 **unique**
usernames a user has. If a user uses the history of [A, B, A, C] only
one instance of A will be logged.

Wrote disabled loop to cleanup usernames in the background. May add it
as a command later.